Fix test issues

+++ application/models/post.php
@@ -28,7 +28,7 @@ class Post extends CI_Model
         $isadmin = (int) $isadmin;
         if($votes < 1) return false;
 
-        $sql = $this->db->query("SELECT id FROM users WHERE email='" . $this->db->escape($email) . "'");
+        $sql = $this->db->query("SELECT id FROM users WHERE email=" . $this->db->escape($email));
 
         if($sql->num_rows()) return false;
 
@@ -99,8 +99,6 @@ class Post extends CI_Model
 
 
     public function vote($idea_id, $user_id, $votes){
-        
-
         $idea_id = (int) $idea_id;
         $user_id = (int) $user_id;
         $votes = (int) $votes;
@@ -155,17 +153,19 @@ class Post extends CI_Model
 
     public function update_by_id($table, $field, $value, $id){
         $id = (int) $id;
-        $table = $this->db->escape($table);
-        $field = $this->db->escape($field);
         $value = $this->db->escape($value);
 
-        $query = "UPDATE $table SET $field='$value' WHERE id='$id'";
+        if(!$this->isAlphaNumeric($table)) return false;
+        if(!$this->isAlphaNumeric($field)) return false;
+
+        $query = "UPDATE $table SET $field=$value WHERE id='$id'";
         $this->db->query($query);
     }
 
     public function delete_row_by_id($table, $id){
         $id = (int) $id;
-        $table = $this->db->escape($table);
+        if(!$this->isAlphaNumeric($table)) return false;
+
 
         $this->db->query("DELETE FROM $table WHERE id='$id'");
     }
@@ -295,6 +295,7 @@ class Post extends CI_Model
 
     private function get_row_by_id($table, $id){
         $id = (int) $id;
+        if(!$this->isAlphaNumeric($table)) return false;
         $sql = $this->db->query("SELECT * FROM $table WHERE id='$id'");
         return $sql->row();
     }
@@ -305,6 +306,9 @@ class Post extends CI_Model
         if(@isset($data->value)) return $data->value;
         else return false;
     }
+
+    private function isAlphaNumeric($text) {
+        return ctype_alnum($text);
+    }
 }
 
-?>
